use strategy pattern to switch ai
