+++ packages/api/src/index.ts
@@ -7,6 +7,7 @@ import * as v from 'valibot'
 import { handleRecommendRequest } from './recommend-handler'
 import type { ErrorResponse } from './types'
 import { RecommendRequestSchema } from './validation'
+import { WorkerAiClient } from './ai/client/worker-ai'
 
 const app = new Hono<{ Bindings: Cloudflare.Env }>()
 
@@ -42,7 +43,7 @@ app.post('/api/recommend', async (c) => {
 
     // AI推論の実行
     const request = parseResult.output
-    const result = await handleRecommendRequest(c.env.AI, request)
+    const result = await handleRecommendRequest(new WorkerAiClient(c.env.AI), request)
 
     if (Result.isFailure(result)) {
       const error = Result.unwrapError(result)
